Create Help
Every good command line tool has a robust "Help" section. This instructs
the user on how to use the tool, what to look out for, gotchas, and
could have some shameful marketing to go along with it.

I want to add a help section to the tool. This help is going to be
minimal for now because my purview for the application at this point is
quite limited. The text in this help would be revised moving forward

This is done by:
- Showing...
    - The syntax for running the application
    - Possible flags one could use to run the application and what these
      flags mean
    - Shameful marketing pointing people to this GitHub account

Resolves #2