Read Configuration
At this point, what Freckle decides to allow is to use one's Personal
Access Token to access the services. This token is generated on and only
on Freckle. I would use this in the app for the time being. I decided to
have this in a configuration file at `~/.bellboy` — pending the app's
name change.

* Did some formating
* Read config file
* Added sample config

Resolves #4