Call me Bellboy
Freckle has some legalities around the naming. It would be unwise to continue
with this as a name for the tool. I would rather have the nanmes that stay
clear of the "Freckle" key word; albeit names like "Freckle Commandline Tool -
FCT" being as descriptive as possible. So I checked what freckles are
associated with, and "Bellboy" came up. ...since this is something which would
be logging my time for me anyways

I did this by:

- changing all `Freckle` to `Bellboy` in any case
- renaming the files as necessary
- ignoring the build executable

Resolves #7