List Projects
The first thing people would need to do is to list the projects which
they can currently log time under. There is a risk in doing this because
there is a rate limiting for 2 requests per second. But I don't think
this should be a problem.

This was done by:

- Reaching the Freckle API using [HTTPoison](https://github.com/edgurgel/httpoison)
- Parsing the resultant JSON using [Poison](https://github.com/devinus/poison)
- Rendering the results as a table with [TableRex](https://github.com/djm/table_rex)

Resolves #10