Fix personal token key name
When trying it out to list my projects after adding the config values, I
was constantly getting an error. After investigations, I found that the app
is trying to get the access token as `freckle_token` instead of
`personal_access_token` which is the correct name according to `config.sample`.
We need to make sure that the app gets the access token from the correct
environment variable.

This change addresses the need by:

* changing `freckle_token` to `personal_access_token`